Dynamically set sensitivity of next button on device selection screen
